section 3 completion refactor
