+++ src/main/definitions/constants.ts
@@ -219,6 +219,7 @@ export const TranslationKeys = {
   CLAIMANT_PLACE_OF_WORK: 'claimant-place-of-work',
   CLAIMANT_RESPONDENT_DETAILS_CHECK: 'claimant-respondent-details-check',
   CLAIMANT_RESPONDENT_ADD_REDIRECT: 'claimant-respondent-add-redirect',
+  CLAIMANT_RESPONDENT_SECTION_CHECK: 'claimant-respondent-section-check',
   CLAIMANT_LINKED_CASES: 'claimant-linked-cases',
   CLAIMANT_CLAIM_TYPE_PAY: 'claimant-claim-type-pay',
   CLAIMANT_COMPENSATION: 'claimant-compensation',
@@ -421,6 +422,7 @@ export const PageUrls = {
   CLAIMANT_PLACE_OF_WORK: '/claimant-place-of-work',
   CLAIMANT_RESPONDENT_DETAILS_CHECK: '/claimant-respondent-details-check',
   CLAIMANT_RESPONDENT_ADD_REDIRECT: '/claimant-respondent-add-redirect',
+  CLAIMANT_RESPONDENT_SECTION_CHECK: '/claimant-respondent-section-check',
   CLAIMANT_LINKED_CASES: '/claimant-linked-cases',
   CLAIMANT_CLAIM_TYPE_PAY: '/claimant-claim-type-pay',
   CLAIMANT_COMPENSATION: '/claimant-compensation',
