+++ src/main/modules/routes/routes.ts
@@ -74,6 +74,7 @@ import ClaimantRespondentDetailsCheckController from '../../controllers/Claimant
 import ClaimantRespondentNameController from '../../controllers/ClaimantRespondentNameController';
 import ClaimantRespondentPostcodeEnterController from '../../controllers/ClaimantRespondentPostcodeEnterController';
 import ClaimantRespondentPostcodeSelectController from '../../controllers/ClaimantRespondentPostcodeSelectController';
+import ClaimantRespondentSectionCheckController from '../../controllers/ClaimantRespondentSectionCheckController';
 import ClaimantTellUsWhatYouWantController from '../../controllers/ClaimantTellUsWhatYouWantController';
 import ClaimantTribunalRecommendationController from '../../controllers/ClaimantTribunalRecommendationController';
 import ClaimantTypeOfClaimController from '../../controllers/ClaimantTypeOfClaimController';
@@ -382,6 +383,8 @@ export class Routes {
     app.get(PageUrls.CLAIMANT_RESPONDENT_ADD_REDIRECT, new ClaimantRespondentAddRedirectController().get);
     app.get(PageUrls.CLAIMANT_RESPONDENT_DETAILS_CHECK, new ClaimantRespondentDetailsCheckController().get);
     app.post(PageUrls.CLAIMANT_RESPONDENT_DETAILS_CHECK, new ClaimantRespondentDetailsCheckController().post);
+    app.get(PageUrls.CLAIMANT_RESPONDENT_SECTION_CHECK, new ClaimantRespondentSectionCheckController().get);
+    app.post(PageUrls.CLAIMANT_RESPONDENT_SECTION_CHECK, new ClaimantRespondentSectionCheckController().post);
     app.get(PageUrls.CLAIMANT_WORK_POSTCODE_ENTER, new ClaimantWorkPostcodeEnterController().get);
     app.post(PageUrls.CLAIMANT_WORK_POSTCODE_ENTER, new ClaimantWorkPostcodeEnterController().post);
     app.get(PageUrls.CLAIMANT_WORK_POSTCODE_SELECT, new ClaimantWorkPostcodeSelectController().get);
