+++ src/main/resources/locales/en/translation/claimant-respondent-section-check.json
@@ -1,0 +1,13 @@
+{
+  "title": "Is this section completed?",
+  "legend": "Have you completed this section?",
+  "yes": "Yes, I've completed this section",
+  "no": "No, I'll come back to it later",
+  "hint": "You can change your answers later.",
+  "errors": {
+    "employmentAndRespondentCheck": {
+      "required": "Tell us if you have completed this section",
+      "invalid": "You must complete the Employment and respondent details section before you can mark it as completed"
+    }
+  }
+}
