+++ src/main/views/claimant-respondent-section-check.njk
@@ -1,0 +1,19 @@
+{% extends "template.njk" %}
+
+{% block beforeContent %}
+  {{ super() }}
+  {% include "back-link.njk"%}
+{% endblock %}
+
+{% block content %}
+  {% block errorSummary %}
+    {% include "error-summary.njk" %}
+  {% endblock %}
+  <div class="govuk-grid-row">
+    <div class="govuk-grid-column-two-thirds">
+      {% block form %}
+        {% include "form/form.njk" %}
+      {% endblock %}
+    </div>
+  </div>
+{% endblock %}
